proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
slowapi==0.1.9
redis==5.0.1 